requirements.txt
fastapi
uvicorn[standard]
python-docx
python-dotenv
pydantic
openai
office365-rest-python-client
requests


